Added methods for creating and deleting playlists.
The whole way of doing mutation calls needs to be cleaned up and handled
in a way that doesn't involve manually building a dict every time. Part
of me thinks it should try to behave like ActiveRecord and the model
should keep track of dirty fields and have methods to generate the
mutation dicts. Another part of me thinks that causes problems with the
current separation of sessions and protocols and data models.